reference/your-podcast-main/backend/pyproject.toml
[project]
name = "backend"
version = "0.2.0"
description = "Your Podcast - FastAPI backend for AI-powered podcast generation"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.0",
    "uvicorn[standard]>=0.41.0",
    # Config
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    # Auth
    "authlib>=1.3",
    "httpx>=0.27",
    "itsdangerous>=2.1",
    # Services (lightweight)
    "feedparser>=6.0",
    "google-genai>=1.0",
    "zhipuai>=2.0",
    "boto3>=1.35",
    "pydub>=0.25",
    # Utilities
    "python-multipart>=0.0.9",
    # Local dev database
    "aiosqlite>=0.20",
    # Schema & migrations
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    # Podcast generation
    "podcastfy>=0.3",
    # podcastfy imports playwright at module level but doesn't declare it
    "playwright>=1.40",
]

[project.optional-dependencies]
test = ["pytest>=8.0", "pytest-asyncio>=0.24", "pytest-cov>=5.0", "httpx>=0.27"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
